requirements.txt
opencv-contrib-python
pillow
scipy
aiofiles
xformers
torch
torchvision
transformers>=4.41.2
diffusers>=0.29.0
safetensors
accelerate>=0.31.0
einops
bitsandbytes==0.39.1
tensorboard
absl-py
ml-collections
datasets==2.21.0
huggingface-hub
peft
sentencepiece
tqdm
waifuset==0.2.11
matplotlib
controlnet_aux
wandb
lion_pytorch